+++ CSGraphQLAPIDeployScripts/websphere/bash/configureGQL.sh
@@ -7,6 +7,8 @@
 # */
 #!/bin/bash
 
+# configureGQLsh
+
 #set -x
 
 #uncomment set -e if want script to continue on error. Else, default to stop at 1st error
@@ -55,6 +57,26 @@ while [ "$1" != "" ]; do
     shift
 done
 
+#set default if value is not set
+if [[ "$GRAPHQL_APP_NAME" == "" ]]; then
+	GRAPHQL_APP_NAME='content-services-graphql'
+fi
+if [[ "$IS_GRAPHQL_COLOCATE_WITH_CPE" == "" ]]; then
+	IS_GRAPHQL_COLOCATE_WITH_CPE="false"
+fi
+
+#skip set up ldap and import CPE LTPA if IS_GRAPHQL_COLOCATE_WITH_CPE = false
+#skip set SSO domain, inbound_realms_trusted since these values are from Global Security and should have been set by CPE
+if [[ "$IS_GRAPHQL_COLOCATE_WITH_CPE" == "true" ]]; then
+	echo "GraphQL is colocated with CPE, script will skip a number of operations"
+	REQUIRE_LDAP_SETUP="false"
+	LTPA_INTERMEDIATE_FILE=""
+	SSO_DOMAIN=""
+        INBOUND_REALMS_TRUSTED=""
+        CPE_SELF_SIGNED_CERT=""
+fi
+
+
 
 if [ "$APPSERVER_BIN_DIR" == "" ]; then
 	echo "APPSERVER_BIN_DIR var is required. Either set it by pass in --file <properties_file> or set shell vars before running."
@@ -74,18 +96,19 @@ echo "#-----------------------"
 date
 if [[ "$REQUIRE_LDAP_SETUP" == "true" ]]; then
 	. ./configureWSLDAPFederated.sh 
+       echo "#-----------------------"
+       echo "#Restart Websphere server "
+       echo "#-----------------------"
+       date
+       . ./restartWAS.sh
 else
 	echo "Skip setting up LDAP federated realm. "
 fi
 
 
-echo "#-----------------------"
-echo "#Restart Websphere server "
-echo "#-----------------------"
-date
-. ./restartWAS.sh
 
-
+# extract this value from CPE appserver > Security > Global Security > Web and SIP Security > SSO > Domain Name
+# since this value is stored in Global Security, if GraphQL and CPE is colocated in the same profile, skip this step
 echo "#-----------------------"
 echo "#enable SSO in GraphQLServer"
 echo "#-----------------------"
@@ -97,14 +120,11 @@ else
 fi
 
 
+#extract LTPA file from CPE appserver > Security > Global Security > LTPA
 #echo
 #echo "#-----------------------"
 #echo "# export LTPA from CEServer "
 #echo "#-----------------------"
-if [[ $LTPA_INTERMEDIATE_FILE == "" || $LTPA_PWD == "" ]]; then
-	echo "Missing defining keys for LTPAImport"
-	exit 1
-fi
 #date
 #./exportLTPAKeys.sh  $LTPA_INTERMEDIATE_FILE $LTPA_PWD  # sanple only
 #echo
@@ -113,22 +133,29 @@ fi
 
 echo
 echo "#-----------------------"
-echo "# import CEServer LTPAKeys from CEServer to GraphQL"
+echo "# import CEServer LTPAKeys from CEServer to GraphQL "
 echo "#-----------------------"
 date
-. ./importLTPAKeys.sh  
-
+if [[ "$LTPA_INTERMEDIATE_FILE" != "" ]] ; then
+   if [[  $LTPA_PWD != "" ]]; then 
+      . ./importLTPAKeys.sh  
+   else
+        echo "Missing defining value for LTPA_PWD"
+   fi
+else
+   echo "Skip importing CPE LTPA "
+fi
 
 echo
 echo "#-----------------------"
 echo "# configure inbound trusted realms"
 echo "#-----------------------"
-if [[ $INBOUND_REALMS_TRUSTED == "" ]]; then
-	echo "Missing defining keys for addTrustedRealm"
-	exit 1
+if [[ $INBOUND_REALMS_TRUSTED != "" ]] ; then
+	date
+        . ./addTrustedRealm.sh
+else
+   echo "Skip configuring inbound trusted realms"
 fi
-date
-. ./addTrustedRealm.sh
 
 echo
 echo "#-----------------------"
@@ -174,13 +201,19 @@ echo "#-----------------------"
 date
 . ./mapSecurity.sh
 
+
+#if CPE is colocated with GraphQL, assume that the Global Security > Application security > Enable application security is checked 
 echo
 echo
 echo "#-----------------------"
 echo "#enable Application Security in GraphQLServer"
 echo "#-----------------------"
 date
-. ./enableAppSecurity.sh
+if [[ "$IS_GRAPHQL_COLOCATE_WITH_CPE" == "false" ]]; then
+   . ./enableAppSecurity.sh
+else
+   echo "Skip enable Global Security > Application security "
+fi
 
 echo 
 echo "#-----------------------"
@@ -197,7 +230,7 @@ if [[ $CPE_SELF_SIGNED_CERT != "" ]]; then
 	date
 	. ./useSSLComm.sh 
 else
-	echo "Skip importing CPE Self signed cert since no CPE_SELF_SIGNED_CERT is specified"
+	echo "Skip importing CPE Self signed cert"
 fi